moved views in seperate folder
