+++ src/app/views/browser/browser.component.html
