+++ src/app/views/browser/browser.component.scss
