+++ src/app/views/browser/browser.component.spec.ts
