+++ src/app/views/color/color.component.html
