+++ src/app/views/color/color.component.scss
