+++ src/app/views/color/color.component.spec.ts
