+++ src/app/views/color/color.component.ts
