+++ src/app/views/editor/editor.component.html
