+++ src/app/views/editor/editor.component.scss
