+++ src/app/views/editor/editor.component.ts
