+++ src/app/views/icons/icons.component.html
