+++ src/app/views/icons/icons.component.scss
