+++ src/app/views/icons/icons.component.spec.ts
