+++ src/app/views/icons/icons.component.ts
