+++ src/app/views/settings/settings.component.ts
